feat: split dashbot models
