+++ lib/dashbot/core/common/pages/dashbot_default_page.dart
@@ -1,7 +1,6 @@
-import 'package:apidash_design_system/apidash_design_system.dart'
-    show kVSpacer20, kVSpacer16, kVSpacer10;
-import '../../../features/chat/models/chat_models.dart';
+import 'package:apidash_design_system/apidash_design_system.dart';
 import '../../../features/home/view/widgets/home_screen_task_button.dart';
+import '../../constants/constants.dart';
 import '../../routes/dashbot_routes.dart';
 import '../../utils/dashbot_icons.dart';
 import 'package:flutter/material.dart';
