+++ lib/dashbot/core/error/chat_failure.dart
@@ -1,0 +1,8 @@
+class ChatFailure implements Exception {
+  final String message;
+  final String? code;
+  const ChatFailure(this.message, {this.code});
+
+  @override
+  String toString() => 'ChatFailure: $message';
+}
