+++ lib/dashbot/core/routes/dashbot_router.dart
@@ -1,6 +1,6 @@
 import '../../features/chat/view/pages/dashbot_chat_page.dart';
-import '../../features/chat/models/chat_models.dart';
 
+import '../constants/constants.dart';
 import 'dashbot_routes.dart';
 import '../common/pages/dashbot_default_page.dart';
 import '../../features/home/view/pages/home_page.dart';
