+++ lib/dashbot/core/services/agent/prompt_builder.dart
@@ -1,7 +1,8 @@
 import 'package:apidash/models/models.dart';
 import 'package:apidash/dashbot/core/constants/dashbot_prompts.dart' as dash;
 
-import '../../../features/chat/models/chat_models.dart';
+import '../../../features/chat/models/chat_message.dart';
+import '../../constants/constants.dart';
 
 class PromptBuilder {
   String buildSystemPrompt(
