+++ lib/dashbot/features/chat/view/widgets/dashbot_task_buttons.dart
@@ -6,7 +6,7 @@ import 'package:apidash/screens/common_widgets/agentic_ui_features/tool_generati
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../models/chat_models.dart';
+import '../../../../core/constants/constants.dart';
 import '../../../home/view/widgets/home_screen_task_button.dart';
 import '../../../../core/providers/dashbot_window_notifier.dart';
